Workaround for Shotgun API bug
When using two levels of redirection in a find request, if the
linked entity used for the first level of redirection is empty,
a CRUD error is issued by SG. For example, "sg_version.Version.entity.Shot.code"
on a cut item will raise such an error if no version is linked to
any of the cut item found by the requests.
The request is now done in two passes, with an additional pass
to reconciliate the data found.