For #30880, handle "repeated" changes on shot name editing
When a shot name is edited, it can become un-repeated or repeated.
When setting the repeated property on a CutDiff, check if the value
changed and emit a signal if it is the case.
Connect repeated changed signal from CutDiff to its widget so
the UI can be refreshed when neeeded.