For #30880, comments and ShotCutDiffList
Added docstring and comments to explain the logic being
cut differences.
Added ShotCutDiffList, a thin wrapper around regular list
so min and max values can be computed when CutDiff instances
are added / removed from a list for a shot.
Added some helpers on CutDiff to retrieve current and new time
code cut in.