For 36807, fixes for shot name editing
Ensured shot name is not editable if the entry is linked to
a Version, either through its CutItem or edit entry.
Fixed a bug when editing shot names, and ensured linked
Version is updated when needed.
Added a "Shot name is not editable" tooltip when the shot name
is not editable.
Added a CutDiff._sg_version member, so linked Version can be edited
without modifying the edit entry or the CutItem.