For #31695, revised total summary
Added a total_count to CutSummary, so the count can be
retrieved from a pre-computed value instead of looping
over all entries. This allows to fix a rare race condition
bug when dealing with lot of entries in an EDL, the
summary being modified while total is counted on another
thread.
Removed CutSummary._rescans_count member and directly use
the _counts dictionary.
Removed useless code.
Only count in total count entries which are not omitted,
that is, entries which do have a CutItem.